organize pomodoro components
- edit pomodoro path in app.js & drawer
